+++ ncrsync/__init__.py
@@ -2,4 +2,4 @@
 using rsync for reliable, resumable downloads.
 """
 
-__version__ = "0.4.1"
+__version__ = "0.5.0"